python_ai_core/cascade/stage1_retrieval.py
import os
import faiss
import numpy as np
from typing import List

# Keep within budget: 35ms embedded FAISS retrieval
# Simulated global index, initialized once on startup
d = 64  # dimension of GraphSAGE embeddings
index = faiss.IndexHNSWFlat(d, 32)
# Dummy logic to add placeholders if index is empty
if index.ntotal == 0:
    index.add(np.random.rand(500, d).astype('float32'))

def faiss_retrieve(user_id: str, current_cart: List[str]) -> List[str]:
    """
    Embedded FAISS HNSW Retrieval replacing Milvus.
    
    Heterogeneous Graph (Users, Items, Restaurants) via GraphSAGE:
    - user_node -> views/buys -> item_node
    - item_node -> belongs_to -> restaurant_node
    
    Item-item similarity computed via Siamese Networks & Triplet Loss:
    L(A, P, N) = max(||f(A) - f(P)||^2 - ||f(A) - f(N)||^2 + margin, 0)
    """
    
    # Query with a dummy vector
    query_vector = np.random.rand(1, d).astype('float32')
    distances, indices = index.search(query_vector, 50)
    
    # Return mapped item IDs (Simulated mapping)
    return [f"I{(i % 300) + 1:04d}" for i in indices[0]]

# PyTorch stubs for the Siamese / GraphSAGE models
import torch
import torch.nn as nn
import torch.nn.functional as F

class SiameseItemNet(nn.Module):
    def __init__(self, emb_dim=64):
        super().__init__()
        self.fc = nn.Sequential(
            nn.Linear(128, 64),
            nn.ReLU(),
            nn.Linear(64, emb_dim)
        )
        
    def forward_one(self, x):
        return self.fc(x)

    def forward(self, anchor, pos, neg):
        a_out = self.forward_one(anchor)
        p_out = self.forward_one(pos)
        n_out = self.forward_one(neg)
        return a_out, p_out, n_out

class TripletLoss(nn.Module):
    def __init__(self, margin=1.0):
        super(TripletLoss, self).__init__()
        self.margin = margin
        
    def forward(self, anchor, positive, negative):
        dist_pos = (anchor - positive).pow(2).sum(1)
        dist_neg = (anchor - negative).pow(2).sum(1)
        loss = F.relu(dist_pos - dist_neg + self.margin)
        return loss.mean()
