python_ai_core/main.py
import os
import time
import grpc
from fastapi import FastAPI, Request, HTTPException
from pydantic import BaseModel
from typing import List

# Stubs for other microservice handlers
from cascade.stage1_retrieval import faiss_retrieve
from cascade.stage2_sequential import bert4rec_score
from cascade.stage3_ranking import pytorch_dlrm_predict
from cascade.component3_multimodal import generate_rag_explanation

app = FastAPI(title="Project Titan CSAO Rail (Lean Dev Mode)")

# L1 Cache (In-Memory)
TOP_50_GLOBAL = ["I0001", "I0002", "I0010"] # Stubs

class InferenceRequest(BaseModel):
    user_id: str
    session_id: str
    cart_items: List[str]
    lat: float
    lon: float

@app.post("/api/v1/recommend")
async def recommend(req: InferenceRequest):
    start_time = time.time()
    
    try:
        # 1. Feature Retrieval (Feast) - Budget 15ms
        user_features = {"time_sin": 0.5, "cart_value": 200}
        
        # 2. Stage 1 Retrieval (FAISS Embedded) - Budget 35ms
        candidates = faiss_retrieve(req.user_id, req.cart_items)
        
        # 3. Stage 2 Sequential (BERT4Rec InMemory)
        seq_candidates = bert4rec_score(req.cart_items, candidates)
        
        # 4. Stage 3 Ranking (DLRM Embedded PyTorch) - Budget 80ms
        ranked_top_50 = pytorch_dlrm_predict(req.user_id, seq_candidates)
        
        # Check latency for Circuit Breaker (Max 250ms)
        elapsed = (time.time() - start_time) * 1000
        if elapsed > 250:
            print(f"Latency Exceeded {elapsed}ms. Triggering Circuit Breaker.")
            return {"status": "circuit_breaker", "recommendations": TOP_50_GLOBAL}
        
        # 5. Component 4 GoLang Re-Ranking (gRPC Stub) - Budget 20ms
        final_top_5 = call_golang_reranker(req, ranked_top_50)
        
        # 6. Component 3 GenAI Explanations (Llama 3 RAG Stub)
        explanation = generate_rag_explanation(final_top_5[0])
        
        return {
            "status": "success",
            "latency_ms": round(elapsed, 2),
            "recommendations": final_top_5,
            "explanation": explanation
        }
        
    except Exception as e:
        # Fallback L2 Redis Cache
        return {"status": "fallback", "recommendations": TOP_50_GLOBAL}

def call_golang_reranker(req, candidates):
    # Stub for gRPC call to rules engine
    # Simulates filtering by Rules Engine constraints
    return candidates[:5]

if __name__ == "__main__":
    import uvicorn
    uvicorn.run(app, host="0.0.0.0", port=8080)
