python_ai_core/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
grpcio==1.56.2
grpcio-tools==1.56.2
pydantic==2.3.0
torch==2.1.0
torchvision==0.16.0
transformers==4.33.1
faiss-cpu==1.7.4
numpy>=1.24.0
